Refactor project; add unit tests for Graph file handler (NOT ALL PASSING yet)
