+++ ReasoningEngine/GraphFileHandling/IndexManager.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 
-namespace GraphFileHandling
+namespace ReasoningEngine.GraphFileHandling
 {
     public class IndexManager
     {
