+++ ReasoningEngine/Utils/DebugUtils/DebugWriter.cs
