--- graph-with-nodes.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "graph-with-nodes", "graph-with-nodes.csproj", "{724C1682-EFE5-460D-9D0C-D23FEC122033}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{724C1682-EFE5-460D-9D0C-D23FEC122033}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{724C1682-EFE5-460D-9D0C-D23FEC122033}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{724C1682-EFE5-460D-9D0C-D23FEC122033}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{724C1682-EFE5-460D-9D0C-D23FEC122033}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {DB561243-59CC-44A1-B998-23FF8BA044E1}
-	EndGlobalSection
-EndGlobal
